[init] init boiler plate
