+++ chrome-extension/public/_locales/en/messages.json
@@ -1,10 +1,10 @@
 {
   "extensionDescription": {
-    "description": "Extension description",
-    "message": "Chrome extension boilerplate developed with Vite, React and Typescript"
+    "description": "auto report",
+    "message": "auto report"
   },
   "extensionName": {
-    "description": "Extension name",
-    "message": "Chrome extension boilerplate"
+    "description": "auto report",
+    "message": "auto report"
   }
 }
